Fixed postmant verify user

+++ backend/job/views.py
@@ -1,12 +1,13 @@
 from django.shortcuts import get_object_or_404, render
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from .models import Job
 from .serializers import JobSerializer
 from rest_framework import status
 from django.db.models import Avg, Count, Min, Max
 from .filters import JobFilter
 from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import IsAuthenticated
 
 # Create your views here.
 
@@ -33,7 +34,9 @@ def getJob(request, pk):
 
 
 @api_view(["POST"])
+@permission_classes([IsAuthenticated])
 def newJob(request):
+    request.data["user"] = request.user
     data = request.data
     job = Job.objects.create(**data)
     serializer = JobSerializer(job, many=False)
@@ -41,8 +44,14 @@ def newJob(request):
 
 
 @api_view(["PUT"])
+@permission_classes([IsAuthenticated])
 def updateJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response(
+            {"message": "You can not update this Job"}, status=status.HTTP_403_FORBIDDEN
+        )
     job.title = request.data["title"]
     job.description = request.data["description"]
     job.email = request.data["email"]
@@ -62,8 +71,14 @@ def updateJob(request, pk):
 
 
 @api_view(["DELETE"])
+@permission_classes([IsAuthenticated])
 def deleteJob(request, pk):
     job = get_object_or_404(Job, id=pk)
+
+    if job.user != request.user:
+        return Response(
+            {"message": "You can not delete this Job"}, status=status.HTTP_403_FORBIDDEN
+        )
 
     job.delete()
